I think am done with CPP06/ex00 Alhamdulillah

+++ CPP06/ex00/Utils.hpp
@@ -15,7 +15,7 @@ bool isFloatLiteral(const std::string &literal);
 bool isDoubleLiteral(const std::string &literal);
 
 void convertFromChar(char c);
-void convertFromInt(int i);
+void convertFromInt(long i);
 void convertFromFloat(float f);
 void convertFromDouble(double d);
